requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.6
PyJWT==2.10.1
pytz==2024.1
sqlparse==0.5.2
drf-writable-nested==0.7.1
